+++ support/scenario-convert.rb
@@ -55,6 +55,39 @@ units(metadata["axis_units"])
 puts <<EOF
     }.freeze
 
+EOF
+
+def init_units(unit_data)
+  unit_data.each do |unit|
+    string = "      { data: :#{unit["data"]["id"]}, x: #{unit["x"]}, y: #{unit["y"]}"
+    string += ", facing: #{unit["facing"]}" if unit["facing"]
+    string += " },"
+    puts string
+  end
+end
+
+if mapdata["init_allied_units"]
+  puts <<EOF
+    INIT_ALLIED_UNITS = [
+EOF
+  init_units(mapdata["init_allied_units"])
+  puts <<EOF
+    ].freeze
+EOF
+end
+
+if mapdata["init_axis_units"]
+  puts <<EOF
+    INIT_AXIS_UNITS = [
+EOF
+  init_units(mapdata["init_axis_units"])
+  puts <<EOF
+    ].freeze
+EOF
+end
+
+puts <<EOF
+
     class << self
       def generate
         {
@@ -69,6 +102,16 @@ puts <<EOF
           allied_units:,
           axis_units:,
 EOF
+if mapdata["init_allied_units"]
+  puts <<EOF
+          init_allied_units:,
+EOF
+end
+if mapdata["init_axis_units"]
+  puts <<EOF
+          init_axis_units:,
+EOF
+end
 if metadata["special_rules"] && metadata["special_rules"].length > 0
   puts <<EOF
           special_rules: #{metadata["special_rules"]},